refactor triplets generation, improve codabase using iterator and binaryheap for performance
