+++ src/observations/mod.rs
@@ -3,3 +3,4 @@ pub mod trajectory_ext;
 mod parquet_reader;
 pub mod ades_reader;
 pub mod observations_ext;
+pub(crate) mod triplets_iod;
